+++ app-backend/migrations/src/main/scala/migrations/Summary.scala
@@ -151,4 +151,5 @@ object MigrationSummary {
   M150
   M151
   M152
+  M153
 }